Move polling for a message into the DockerLogsContainer task
